Change databag schema, some renaming and formatting
